Working v1 /report endpoint; not displaying data yet.
Server contains two permanent endpoints (GET / and POST /report) that
are used for viewing data and storing it. There's also a temporary
endpoint (GET /init) that can be used to initialize a new database
for testing purposes.

Note that very little validation is being done and no authentication.
This is *not* production ready by any stretch of the imagination :)